Small change to AC sensitivity std output format
While working on this issue:
"Expression-based objective functions for .SENS conflate I(VA) with V(VA) (nodes and devices w/ same names)"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/729

I noticed that the formatted std output for AC sensitivities
didn't include the frequency for each output.  As this is obviously
a bit of information a user would want, this commit adds it.