Fix the spice strategy in nonlinear solver
The spice strategy is the default nonlinear solver for DC
calculation. However, if a DC solve fails during a DC sweep, all
the solutions after that can be incorrect with default spice
strategy. If a DC solve fails, it fails at the last step of
spice strategy which is source stepping. The scaling factor
used in source stepping was not reset correctly. For the next
DC solve, the simulation problem was based on the scaled sources
from the last failed step. This can lead to incorrect results after
a failed DC solve. This commit fixes the issue.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/722