Correct TEMP param mistake from previous commit
In my previous commit, I addressed this issue:

"Implement DTEMP parameter for R, L, C, D, Q, M and J devices"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/343

The mistake was that I changed the "addPar" function call for the TEMP
parameter in each of these devices.  I changed the default value from
0.0 to the actual default number.  In practice, this change wouldn't
mattter much, as the default gets set outside the parameter handling
machinery anyway, but it seemed reasonable to make it the actual default
to avoid confusion.

However, I had forgotten that setting TEMP to a default value of 0.0
causes the "Xyce -doc" capability to automatically put "Ambient
Temperature" as the default value in the latex file parameter table for
this device.  So, making this change broke the DEV_DOC and DEV_DOC_CAT
tests, in a way I had not intended.

This commit restores the default to the mysterious 0.0 value.