Update reference guide for DTEMP
This is part of my work on this issue:
"Implement DTEMP parameter for R, L, C, D, Q, M and J devices"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/343

Several device models previously did not support DTEMP and now they do.
This commit adds DTEMP to their parameter tables in the reference guide.

Also, this commit contains some minor updates to the addPar functions,
which determine some of the details of the latex outputs.