Fixes to MESFET for DTEMP
Testing of DTEMP on the MESFET revealed that it wasn't set up
correctly.  Some of the dtemp-related code needed to be in the
"processParams" function but it was not.

The MESFET and JFET are nearly identical, so I tracked this down by
performing diffs on the two models.  In addition to the bug fix in
this commit, it also contains a number of cosmetic changes to
the MESFET and JFET so that they match better.

This is part of my work on this issue:
"Implement DTEMP parameter for R, L, C, D, Q, M and J devices"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/343