Remove use of Epetra wrapper methods for MPI
Most of the methods in the Xyce's parallel communicator class are
implemented with MPI calls. There are a handful that used an epetra
communicator. Epetra is being deprecated, so the functions in Xyce's
communicator that use epetra communicator need to be refactored. In
these cases, the epetra communicator methods are just lite wrappers for
MPI. For ease of maintenance and consistency with the rest of the
communicator implementation, the Xyce communicator methods that use the
epetra communicator have been refactored to use MPI directly.

The epetra communicator object is still part of the Xyce communicator,
but transitioning that to another supported communicator in Trilonos is
another issue.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/work_items/801