Add support for .param-based functions
This is for this issue:
"Modify the Xyce parser to handle Hspice-style functions (.param rather
than .func)"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/488

The changes for this commit are mostly in the N_IO_ParsingHelpers files,
in the function IO::extractParamData and related support functions.
This function is called from the circuit block class, so that also
needed modification.  The context doesn't have any meaningful changes in
this commit, just some updates to a few comments.