Remove creation of petraComm_ in Parallel::GlobalAccessor
The Parallel::GlobalAccessor contains a Parallel::Communicator object
that also has an Epetra::Comm object.  Instead of using that communicator,
the Parallel::GlobalAccessor class is creating a new Epetra::Comm object.
The logic to create that communicator is a copied pattern that controls
serial execution if the parallel binary is run on one processor.  That
logic should be contained in as few places as possible in the event that
it needs to change.  So, the Parallel::GlobalAccessor should really get
the Epetra::Comm object from the Parallel::Communicator.  This commit
makes that change.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/810