Add more DTEMP code to diode and bsim3
This is part of my work on this issue:
"Implement DTEMP parameter for R, L, C, D, Q, M and J devices"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/343

DTEMP was added as a parameter to the diode and the BSIM3 models
recently.  However, when I did this work I overlooked the extra code in
these models that supports .SENS.  Templated versions of functions such
as processParams and updateTemperature (among others) were created to
enable the use of automatic differentiation via Sacado FAD types.  These
functions are NOT used in normal device operation.  They are only used
for .SENS.

This commit adds new DTEMP code to the .SENS functions.