Update version string in ADMS back-end to 7.9
The ADMS devices are about to be regenerated because of changes to the
back end of ADMS.   Most of the back end changes happened a few months
ago with the work to implement the BSIM CMG 111.2.1 device.

But, now with this version string change, there is another reason to do
it.  The current devices in the DeviceModelPKG/ADMS directory all say
version 7.7, so that should be fixed.