Update exp unit test Makefiles for c++17, trilinos
The primary way to build the expression library unit tests is using the
CMake build system.  However, at the moment this process appears to be
broken.  I've fixed at least one issue preventing CMake from building
the unit tests on my laptop, but I haven't managed to fix everything.

I find that I'm still able to build the unit test programs, "parserUnitTest"
and "complexParserUnitTest" using my original crude, hand-written makefiles.
And of course, as the makefiles are very simple, it is much easier to
debug them, sot that is what this commit is about.

When I originally wrote these makefiles, I was using the very old
Trilinos version 12.12.   I mostly compile Xyce using a more recent
version of trilinos these days, and the makefiles needed small udpates
to support this.  Mainly changing the c++11 flag to c++17.
This commit accomplishes that.

Regarding the CMake build, I removed the specific required version number for
Trilinos in the CMaskLists.txt directory.  It was set to 13.5, but I've been
using a more recent version of trilinos than that, which caused the
CMake build to break.   The version of trilinos I've been using seems to work
just fine, as evidenced by the crude makefiles.  So removing this
specific requirement seems fine.  That issue (trilinos version)  was the first
problem preventing the CMake build of these tests from working on my laptop.
There are other issues, however, unfixed in this commit.

Also, while I was at it, I removed the hand-written Makefile.bsd from
this directory, as it has not been maintained at all, and I don't have
a BSD platform to test it on.