Add support for M= multipliers in Bsrc and VCCS
This is to fix this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/831
"Add multiplier (M=) to the nonlinear VCCS device (ie, the Bsrc) to support
implicit multilpiers"

Note, VCCS = voltage controlled current source.

I added support for "M=" multipliers in the linear VCCS device (or
Gsource)  a while ago.  However, when the Gsource in nonlinear, and
depends on an expression, the Xyce parser turns it into a Bsrc
internally.  The Bsrc did not support multipliers, so neither did
nonlinear VCCS devices.

This commit adds code to the Bsrc to support multipliers, for the
current source version of Bsrc's.   For the voltage source version,
multipliers don't make sense, so they are not applied in that case.

This commit also contains fixes in the parser to ensure that G-source
parsing does the following:

- VCCS devices properly parses the "M" parameter.
- can be invoked inside of subcircuits, when the solution-dependent
  expression depends on local subcircuit nodes.
- works correctly for both explicit and implicit (i.e. from subcircuit
  M=) multipliers

The above issues were broken, due to the special way nonlinear VCCS
devices are handled in the parser and the very fragile code that
implements this special case.   As noted above, VCCS devices are converted
into B-sources internally.  This conversion happens inside the
DeviceBlock::extractBehavioralDeviceData function.  The subcircuit node
aliasing (if necessary) is handled later in pass-2, in the
DistToolBase::instantiateDevice function.

Both of these functions require a list of device prefixes ("B", "G",
etc) to know if they should be processed or not.  The first one
(from the DeviceBlock) also uses a list of non-expression parameters,
so that they don't merge them together in the expression.

For example, if I had this text on a G-source netlist line: "CUR={expression} M=10",
then before this fix, the parser would incorrectly try to use the entire
"{expression}M=10" as the expression.  As written, the code currently
checks for other valid parameters on the line (like M) and terminates the
expression string at the prior field.  But, this shouldn't require a
list like this.

For this commit, I haven't changed this overall fragile structure, but
I've made the nonlinear G-source work correctly, by adding M= to the
parameter list,  and also adding "G" to various if-statements in both
fucntions.

Finally, I also double check to ensure that the converted
G-source doesn't add the "I" or "V" parameters more than once.  It was
doing this before this commit, which confused the downstream code.