Fix implicit M= multipliers nonlinear VCCS devices
This is to fix this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/831
"Add multiplier (M=) to the nonlinear VCCS device (ie, the Bsrc) to support
implicit multilpiers"

In my previos commit, I mostly got multipliers set up correctly for the
nonlinear VCCS device but not quite.  The one use case that wasn't
set up correctly was the purely implicit case.  i.e. prior to this
commit this worked:

Xtest 3 0 2 0 gsrcSub m=10

.subckt gsrcSub A B C D M=1
G A B cur='v(C,D)*0.02' M='M'
.ends

But, this did not work:

Xtest 3 0 2 0 gsrcSub m=10

.subckt gsrcSub A B C D M=1
G A B cur='v(C,D)*0.02'
.ends

In other words, if the "G" device inside the subcircuit didn't have an
explicit "M" parameter on its line, then the "m=10" from the "X" line
did not get applied to it.   However, given how implicit multipliers
work, it should have been applied.

As with most of the other issues for this device, the issue was in the
parser, and its awkward conversion (internally) from a nonlinear VCCS
device to a Bsrc.   When that conversion happens, it throws away the
parameter metadata for VCCS and/or Bsrc devices.  So, the list of device
parameters after this stage no longer includes "M", even though it is
defined as a supported parameter for both "G" and "B" devices.

With the "M" parameter missing from the parameter metadata, the code
that applies the implicit subcircuit "M" wasn't able to find it, and
thus could not apply it.

The fix for now was to add code to ensure that "M" is always in the
converted device metadata.  Ultimately, it shouldn't be throwing away
the (known) metadata at all, but that fix will happen some other time.