Add documentation of VCCS/Bsrc multipliers to docs
This commit adds documentation for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/831
"Add multiplier (M=) to the nonlinear VCCS device (ie, the Bsrc) to support
implicit multilpiers"

Two bullets are now included in the "new features" section of the release
notes; one for G-sources (VCCS) and one for B-sources.

I have also updated the B-source section reference guide, to include
mention of M= multipliers, and now (for the first time) a parameter
table.

The VCCS device was already documented as supporting multipliers, so
those sections didn't need any modification.