Fix Intel compiler options for Windows
When using the Intel OneAPI compilers to build Xyce, the floating-point
model needs to be set to precise (otherwise certain test/use cases may
fail). A check was added to the CMake files to include the
floating-point model option if the Intel compiler is detected. However,
the check only inserts the fp option that is compatible with the Linux
version of the Intel compiler. This commit expands the logic to also
check the platform, and to use the appropriate form of the fp option.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/834