Set up CMake files to add Xyce_Regression
This sets up cmake to include Xyce_Regression under
Xyce/test as Xyce/test/Xyce_Regression.  Like the addition
of optional source directories, the Xyce_Regression directory
is added if it exists at the same directory level as Xyce.  The
location of Xyce_Regression can also be set with the variable
Xyce_REGRESSION_DIR as in cmake -DXyce_REGRESSION_DIR="a/b/c"

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/658