Use tolerances in more numeric comparisons
These unit tests compute numeric values using different methods
and then used EXPECT_EQ() to test the results.  On some platforms
the values are very close but not exactly equal.  Switch to useing
EXPECT_NEAR() with a tolerance of 5e-13 which is just above some
of the errors found on Arm64.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/658