Fix mishandling of second derivatives
A long time ago, when I implemented ddx in Xyce/ADMS, I made a huge
error in handling of second derivatives that caused massive bloat in
codes that didn't even need them.  In fixing that, I removed all that
bloat but also managed to disable generation of some second derivative
variable declaration when it *WAS* needed.

Second derivatives are absolutely required when any variable that is
assigned a value using "ddx" is then used in any computation that ends
up in a contribution.  That's because the derivatives needed for the
jacobian of any expression that is already a derivative are in fact
second derivatives now.

I discovered this today when attempting to build the ekv3 version 302
code that Matthias Bucher has released under a much less restrictive
license than the Xyce team was forced to obey years ago when we signed
the NDA for it.

The new version of ekv3 has some output variables that use ddx, but
somehow also manage to get into contributions.

The changes here accomplish several things:
  - Set the correctly named "doSecondDerivs" variable in
  xyceImplementationFile_nosac.xml instead of the variable
  "doSecondDerivatives" (the former is actually used by
  xyceBasicTemplates_nosac, and the latter is never, ever used).
  - Fix the logic in xyceImplementationFile_nosac so that the variable
  is actually set when it needs to be, which is when *any* variable
  must have second derivatives.  Previously it only recognized when
  top level variables needed it.
  - add code to initialize second derivative variables when they
  happen to be instance variables.
  - make sure that all Makefiles in the ADMS directory call the
  xyceHeaderFile_nosac.xml file *after*
  xyceImplementationFile_nosac.xml, because it is the latter file that
  sets "doSecondDerivs", but the former file needs this set properly.
  - change both versions of buildxyceplugin to get the order of
  header/implementation right.

With these changes I can regenerate all of the current ADMS
devices (using "generate_ADMS.sh") and see NO differences.  (Well, no
differences that are caused by this change.  There are differences
because the devices were not regenerated after the ternary operator
fix, but those are good differences.)

And with these changes, I can use buildxyceplugin to generate a plugin
for the ekv3 that compiles Dietmar Warning's fork of the ekv3model
repo.  I have NOT attempted to *run* that model yet, but it does
compile into a plugin with these changes.  It puked with undeclared
variable problems before.