Clean up admsXml command lines
When I fixed the ordering of XML execution, I did it by cutting and
pasting lines into different order.  That led to the actual argument
that contained the Verilog-A file name being in between two "-e"
arguments.

That actually worked just fine, but looks wrong and so I'm fixing it.

GH-104