Fix xyceHeaderFile second deriv logic
When I fixed xyceImplementationFile's setting of "doSecondDerivs" to
fix the logic and actually make it set a variable that is ever used, I
did not notice that xyceHeaderFile had a similar broken block.

I "fixed" things by making sure that all generation of models in the
examples directory or by buildxyceplugin called
xyceImplementationFile_nosac.xml first (thereby getting the correct
variable set properly) and not fixing xyceHeaderFile_nosac at the same
time.

This commit copies the relevant code from xyceImplementationFile_nosac
into xyceHeaderFile_nosac, and removes the old broken code (that
amounted to a no-op, since it set a variable that nobody uses
anywhere).

Doing this now makes it possible to call the two headers in any order,
which is what I should have done in the first place.