Remove post build command to copy of mkl dll
When building xyce with cmake on Windows, there is a post-build command
that copies an mkl dll into the build folder (mkl_sequential.2.dll). I
don't think this copy command is necessary anymore. I think the current
windows configuration is statically linking mkl. The mkl_sequential dll
definitely not packaged into the installer, so it looks like it's just
being copied to the build folder.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/846