Fix for direct sensitivities with PWL source.
This is related to this issue:
"Sensitivities don't work with PWL source"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/849

This came up in an interaction with an external Xyce user.  The user
wanted to compute sensitivities for the values of a PWL vsrc source.  In the
current code configuration, the only way to do this is to set the PWL
values using .param, and then having the relevant .param be the
sensitivity parameters.

In this case, the device derivatives (the df/dp term in the sensitivity
eqaution) must be computed using a finite difference, which is not ideal
but it works.

This should have worked, at least for direct sensitivities. Some uses of
finite difference df/dp don't work for transient adjoints, but they
generally work for direct methods.

It did work for the Bsrc equivalent, via the expression library.
But it didn't work in the vsrc (or probably isrc) because the
"PWLinData::setParams" function did not call "updateSource after
updating the relevant parameters.   This commit fixes that oversight.