Update path for Intel runtime libraries (Windows)
We’re working on updating the Intel oneAPI toolkit for building Xyce
on Windows. The new (2024) edition of oneAPI supports a new directory
structure, the unified directory layout. There are several runtime
dependencies in the toolkit that need to be packaged into the Xyce
installer. The instructions for finding and including these
dependencies are in the cpack configuration, and only work with the
older oneAPI component directory layout. This commit updates the search
paths for the oneAPI runtime dependencies to support both directory
layout models.

Additionally, the cpack operations to copy/package files don’t appear
to work with symbolic links on Windows (and the new oneAPI search paths
includes symbolic links to the library files). So we’re also making
sure that the found library paths are real, absolute paths.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/854