Add "getCoeffs"  function to several interpolators.
For one of my models, which uses splines, it is useful to be able to get
the interpolation coefficients for the cubic polynomial of a specified
cell.  This commit includes implementations of a new "getCoeffs"
function for several of the cubic splines (cubic, akima, wodicka and
steffen).

Setting these up for akima, wodicka and steffen was trivial, as they set
up and store the coefficients in a standard polynomial form.  ie,
y = p0 + p1*x + p2*x*x + p3*x*x*x.

However, the cubic spline implementation here doesn't do this, and uses
a different form to support the interpolation.  Thus, setting up the
function for cubic splines required a bunch of algebraic manipulations.
Fortunately, cubic splines are not the main method used by the model,
for various reasons.