Fix the inf/NaN issue in diode current calculation
The saturation current of zero leads to divide by zero in an internal
variable calculation in the didoe model. The internal variable is
used to set up the initial junction voltage of diode. This leads to
Inf and NaN in diode current calculation and convergence issues.
This commit fixes it. A warning is also outputted for 0 saturation current

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/861