Improve ZONE field in tecplot output
When using format=tecplot, one of the fields is called "ZONE".

ZONES in tecplot are separate datasets.  If running with .STEP, each
STEP iteration outputs a unique zone, with a zone title that describes
the parameter being swept.

When running a non-STEP simulation, ZONE just defaults to the title
"Xyce data", which isn't informative.  Also, sometimes I load a bunch of
separate tecplot files, and each is stored internally in tecplot as a
different ZONE.  However, if they are non-STEP data, then they all have
the same ZONE name, which is confusing.

So, to address this usecase, I chose to make the zone
title match the netlist filename.   That way if I load a bunch of
files, they have zone names matching their original filenames.

This commit also includes very minor edits to the level=1 BJT