Correct version check for cmake policy CMP0144
This commit updates the cmake version check associated with enabling
cmake policy CMP0144.

In a previous merge request, a new cmake policy was enabled to reduce
the number of warnings reported when using newer versions of cmake
(from 3.27 or greater).
https://cee-gitlab.sandia.gov/Xyce/code/uur-proprietary/Xyce/-/merge_requests/87

To avoid raising the overall minimum version required to build Xyce,
the policy was enabled inside a conditional block, but the condition
checks for the wrong version of cmake (the check is for 3.26.0, but
CMP0144 was introduced in 3.27). The result of the mismatch is that
attempting to build using cmake 3.26.x would report an unknown policy
error and fail to configure.