Fix bad memory access in packed restart
For a very long time (longer than git public history), the DeviceMgr
function "dumpRestartData" and "restoreRestartData" have had an error
in dumping "ltraTimePoints_" --- both unconditionally access the
pointer to ltraTimePoints_[0] even if the vector is empty.  This is a
recipe for undefined behavior.

Undefined behavior in this case means "always worked before up to and
including FreeBSD 13.3 but suddenly throws illegal instruction errors
in FreeBSD 13.4, presumably due to a Clang update."  Apparently, that
pointer is invalid when the vector has zero size, and now clang pukes.

Because unpacked restart loops over the size of ltraTimePoints_, it
does nothing to the vector if size is zero.

Several packed restart test cases started failing on my system because
of this error.

This commit replaces several uses of &(vector[0]) with vector.data(),
which is the safer way to do what is trying to be done, and does the
right thing when the vector is empty (provided one doesn't try to
dereference and index off of the pointer received, of course).

This fix is exactly the same as the one applied last February in
Github commit 92e522bf5.  That commit explicitly said it didn't fix
all of this sort of error, just the few that got spotted with respect
to certain internal bug reports.  This commit fixes a few more of
them, and makes no attempt to fix them all.