Fix illegal instruction errors on FreeBSD with clang
On FreeBSD 13.4 with clang 18.1.6, parallel Xyce has started crashing
on over 1900 netlists in the test suite.  A previously built binary
from FreeBSD 13.3 (with Clang 17.something) still runs fine.

These failures are almost all due to attempts to access a pointer to
the underlying storage of an STL vector, without checking that the
vector isn't empty.  This introduces a presumption that the STL
implementation always allocates some storage even if the vector is
empty, which is clearly NOT the case beginning with FreeBSD 13.4 with
Clang 18.

The code in each of these cases looks something like trying to pass
&(vector[0]) to a function that expects a double pointer.  The result
on BSD is that the code exists with an illegal instruction error
inside an assertion that is doing vector "operator[]" bounds checking.

Replacing each of these `&(vector[0])` usages with `vector.data()`
bandaids the problem --- it "fixes the glitch" in the sense that
`data()` doesn't crash, but it only works because in fact the
downstream code doesn't actually try to dereference the pointer it
receives.

I should note in passing that ".data()" was not an option back in the
day when most of this code was written, it was added after C++98.
Back then, the *right* thing to have done would have been to check the
size of the vector before trying to access its underlying storage and
not to try to get it if there was nothing stored.  That would still be
a reasonable option today, but I opted for the more compact, simpler fix.

`data()` is NOT guaranteed to return a non-NULL pointer if
the vector is empty, but at least it doesn't crash.   It remains
unsafe to dereference the pointer, but that's not a problem in any of
these failures --- the pointer is basically never used after obtaining
it, because other indications of the emptiness of the vector are
protecting it.

However a very large number of test failures are due to a strange use
in two spots in N_PDS_MPI.C of the std::string constructor being
called with two char pointer arguments,

    somestring = std::string(&receive[offset], &receive[offset + count]);

presumably intended to mean "copy the string from this pointer up to
that one", but this usage is not apparently part of the standard and I
can find no documentation that it is legal.   Replacing one such
instance of this usage with

      somestring = std::string(&receive[offset], count);

fixes a very large number of failures (some 900 of them).  Fixing a
second occurrance of this same thing fixes another 46.

This single commit is actually a squash of 7 commits I actually made
in addressing this issue, each fixing a subset of tests en masse, but
the details and extensive commentary I made in that process are
uninteresting.  Applying this single squashed commit fixes almost all
of my parallel test failures.  GH-114 fixes a few additional restart
tests that failed in serial for the same reason, and both that PR and
this need to be applied to fix both GH-114 and GH-115.

A small number of failures are actually due to exactly the same issue
in a couple of places inside Trilinos.  See issue GH-115 for more
details on that.

I have ONLY fixed these `&vector[0]` things that actually get executed
during the course of the test suite.  I know that many more are
present throughout Xyce, but clearly aren't being covered during a run
of the test suite so I left them alone.  I consider them undefused
landmines that are going to take someone's leg off someday, so the
Xyce team should probably go through the code carefully and fix the
rest of them.

The team might also consider whether it's a good idea to be peeking
behind the encapsulation of STL vectors so often --- is it always
necessary to get a pointer to the raw underlying storage (obviously,
in some cases it may be necessary for efficiency's sake).

This commit closes GH-115.