Fix 2d array construction in python interface
The python interface was using the notation [[0]*col]*rows to
make a 2d array but this makes one vector with each row
referencing the same vector.  That leads to update problems.
The right way is to use the syntax
[[0 for j in range(col)] for i in range(row)]

With that fixed the python layer now returns the right data
from the ADC's in Xyce.  Added in a Python example that tests
this functionality

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/864