Update Simulink-Xyce control gui to simplify port assignment
Update the Simulink-Xyce control gui to make it easier to add
and subtract rows to the input and output ports tables.  Updated
the examples as well and provide Matlab 2023a compatible versions
of the files.  I didn't make the original versions 2023a because
as soon as they are saved from a later version of Matlab they will
be saved in the newer format.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/864