Allow TABLE to be non-monotonic
This commit adds some sorting routines to the tableOp class.

This allows all the interpolation-based operators to have a
non-monotonic specification.  That includes SPLINE, AKIMA, BARYCENTRIC,
etc.

This only works for purely numerical tables.   For tables that contain
entries dependent on mutable .params, it isn't possible to do this
safely; what might be monotonic for one set of .param values may not be
monotonic for another set.  So the sorting is only performed when the
values are all constant numbers.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/876