Fix odd number of sample point tests.
In testing the underlying FFT interface it is important to
test with both a power of 2 number of sample points and something
other an a power of 2.  Some FFT kernals will behave differently
with a non-power of 2 input.   This commit gets the non power
of 2 tests working with the function samples, sine, gaussian
and step of various frequency.