Copy input test files to build directory
The FFTUnitTests used several signals as input for testing.
The test property WORKING_DIRECTORY is set to run in the
source directory where these files are located but this
doesn't work when the test is just run from the build directory.
Copying the input files to the build directory allows the
test to run properly there too.