Replace nonstandard ASSERT with standard assert
KSparse uses a non-standard macro called ASSERT to accomplish what the
C standard assert() function is supposed to do.

In ngspice they've replaced all that with standard assert calls.

In Xyce, I'm finding that all runs I've tried with ksparse are
crashing with abort traps in these macro calls, even though when
viewed in the debugger the condition appears to be true (it should
abort only if false).

When I replace these nonstandard macro calls with standard C function
calls, no assertion failure is detected.  So clearly there is
something wrong with how the macro is written, but rather than fix
that kludge with further kludges, I'm replacing them all with assert
calls and calling it done.

I have tested this code on the netlists in GH-123 and GH-124, both of
which were crashing this way.  After the fix, all those netlists run
with "-linsolv ksparse" just fine.

GH-123
GH-124
Closes GH-125