Reset root of master branch
Starting a new version control management process. To facilitate the new
process, it was easier to break the repository history than to work with
disjoint histories. To maintain a comprehensive account of the project's
history, the new root node is based on the head of the old line of
history. All future updates will be pushed to branches rooted in this
commit.