Apply fix to BSIM4.7 CGS and CGD output variable
A user reported a dramatic error between Xyce's output of CGS for the
BSIM4.7 model and that of a commercial simulator.  I found that I
could reproduce that easily using one of the simplest netlists in the
regression suite (BSIM4_v4p7/test1.cir).  I tracked the differences
down to Xyce's setting of ChargeComputationNeeded and the choice of
which variable to store for output (Xyce stores CAPcgsb  instead of
cgsb).

The first one meant that cgsb was always zero for DC sweeps, and the
latter meant that the wrong variable was being output.

This commit fixes this issue for CGS and also for CGD (which had the
same mistake) for all three versions of BSIM4.   BSIM3 has the same
problem, but that is not addressed here.

I have NOT fixed the issue that none of the output variables are
appropriately scaled by the multiplicity.  That would be an issue that
touches too many devices for a simple PR like this one.

Closes GH-128