Add calculation routines for unsupported sample sizes
MacOS FFT and IFT routines only support powers of two for the
number of sample points.  Padding a non power of two input with
zeros will work but the result doesn't agree with what FFTW does.
As Xyce doesn't limit the use of FFTs to powers of 2 and its better
to support that FFTW does when using a non-FFTW engine.  So I've
added fall back calculations for FFT and IFT to the AppleFFT
interface to handle sample sizes not normally supported that
do the same that FFTW does.

https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/40