Fix for 1D doping interpolations
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/884

I was working with one of the rad-aware PDE (TCAD) device models
recently, and noticed that under at least one circumstance the spline
interpolations of file-specified doping profiles was producing
oscillating spikes.  This was exhibitied in at least one test case, but
as the test case doesn't specifically look at mesh quantities it wasn't
caught.  That test only looks at terminal currents.

One such test is the PDE_1D_BJT/npn_bjt_pde.cir test.  It has pretty
high doping in the emitter, and the test uses a feature "nmaxchop" which
tells the simulator to truncate doping values that exceed the chop
value.   This parameter was not being handled well; it was applied after
the spline was initialized, so it wasn't consistent.   That has been
fixed.

Also, I've modified the doping spline interpolations so that they
convert the doping to a log scale first.  This (may) also help with the
behavior.

With these changes the spikey behavior in the doping profiles goes away.