Reorder includes to pick up system math constants
The macro M_PI is used in a couple of the unit tests that were added for
this issue. In our Windows build, this macro was not being resolved
correctly. In one case, the definition was not found, and in the other it
was getting the N_UTL_Math.h. In both cases, it should be getting the
value from a system header file.

This commit changes the list of include files in the unit test files so
that, on Windows, the value for M_PI is coming from a system header file
definition .

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/864