Fix for raw output overrides with adjoints
This is a fix for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/888

It was observed by an external user.  The fix is very easy, and involves
a single line.

THe problem was this: when using the command line override for raw files
(via the "-r" option), Xyce attempts to put the contents of all the
netlist .PRINT lines into the raw file.  However, this doesn't work for
unsupported analyses like sensitivities.   What *should* have happened
was that the adjoint information was excluded from the raw file. But, it
was incorrectly appended anyway, in the text-based STD format.

This comit fixes the issue.